moved the damage to entity package
